feat: replace teamStatusWindow init to renderer process
