+++ app/main/sagas/initialize.js
@@ -4,10 +4,8 @@ import {
   Tray,
   Menu,
   MenuItem,
-  screen,
 } from 'electron';
 import path from 'path';
-import Positioner from 'electron-positioner';
 
 import {
   take,
@@ -159,19 +157,6 @@ function* trayManager() {
   }
 }
 
-function teamStatusListTrayManager() {
-  function getIconByName(name) {
-    if (process.env.NODE_ENV === 'development') {
-      return path.join(__dirname, `../../renderer/assets/images/${name}.png`);
-    }
-    return path.join(__dirname, `./app/renderer/assets/images/${name}.png`);
-  }
-
-  const tray = new Tray(getIconByName('icon'));
-  tray.setToolTip('Chronos Team Status');
-  global.teamStatusListTray = tray;
-}
-
 function* onClose({
   win,
   channel,
@@ -189,113 +174,13 @@ function* onClose({
   }
 }
 
-function getWindowPosition(window) {
-  const windowPositioner = new Positioner(window);
-  const screenSize = screen.getPrimaryDisplay().workAreaSize;
-  const windowSize = window.getBounds();
-  const trayBounds = global.teamStatusListTray.getBounds();
-
-  const windowLength = (trayBounds.x + (trayBounds.width / 2)) + windowSize.width;
-  const windowOutOfBounds = (windowLength > screenSize.width);
-
-  const halfScreenWidth = screenSize.width / 2;
-  const halfScreenHeight = screenSize.height / 2;
-
-  const calculateWindowPosition = trayPosition => (offset) => {
-    const { x, y } = windowPositioner.calculate(trayPosition, trayBounds);
-    return { x: x + offset.x, y: y + offset.y };
-  };
-
-  const calculateOffsetCenter = calculateWindowPosition('trayCenter');
-  const calculateOffsetBottomLeft = calculateWindowPosition('trayBottomLeft');
-  const calculateOffsetBottomCenter = calculateWindowPosition('trayBottomCenter');
-  const calculateOffsetTopRight = calculateWindowPosition('topRight');
-
-  if (process.platform === 'win32') {
-    // Vertical or Horizontal Taskbar
-    if ((trayBounds.x + trayBounds.width) <= halfScreenWidth) {
-      // left
-      if (windowOutOfBounds) {
-        return calculateOffsetBottomLeft({ x: -7, y: 0 });
-      }
-      return calculateOffsetBottomLeft({ x: 78, y: -10 });
-    }
-
-    if ((trayBounds.x + trayBounds.width) >= halfScreenWidth) {
-      if ((trayBounds.y + trayBounds.height) <= halfScreenHeight) {
-        // top
-        if (windowOutOfBounds) {
-          return calculateOffsetCenter({ x: -7, y: 0 });
-        }
-        return calculateOffsetCenter({ x: 0, y: 8 });
-      }
-      // bottom or right
-      if (windowOutOfBounds) {
-        return calculateOffsetBottomCenter({ x: -7, y: 0 });
-      }
-      return calculateOffsetBottomCenter({ x: 0, y: -6 });
-    }
-  }
-  if (process.platform === 'darwin') {
-    if (windowOutOfBounds) {
-      return calculateOffsetCenter({ x: -20, y: 0 });
-    }
-    return calculateOffsetCenter({ x: 0, y: 3 });
-  }
-
-  if (windowOutOfBounds) {
-    return calculateOffsetTopRight({ x: -20, y: 0 });
-  }
-  return calculateOffsetTopRight({ x: -10, y: 5 });
-}
-
-function showWindow(window) {
-  const position = getWindowPosition(window);
-  window.setPosition(position.x, position.y, !!process.platform === 'darwin');
-  window.show();
-  window.focus();
-}
-
-function* onToggleClick({
-  win,
-  channel,
-}) {
-  while (true) {
-    yield take(channel);
-    if (win.isVisible()) {
-      global.teamStatusListTray.setHighlightMode('never');
-      win.hide();
-    } else {
-      global.teamStatusListTray.setHighlightMode('always');
-      yield call(showWindow, win);
-    }
-  }
-}
-
-function* onBlur({
-  win,
-  channel,
-}) {
-  while (true) {
-    yield take(channel);
-    global.teamStatusListTray.setHighlightMode('never');
-    win.hide();
-  }
-}
-
 function* forkInitialRendererProcess() {
   try {
     yield fork(trayManager);
-    yield call(teamStatusListTrayManager);
     const url = (
       process.env.NODE_ENV === 'development'
         ? 'http://localhost:3000'
         : `file://${__dirname}/index.html`
-    );
-    const teamStatusListUrl = (
-      process.env.NODE_ENV === 'development'
-        ? 'http://localhost:3000/teamStatusList.html'
-        : `file://${__dirname}../teamStatusList.html`
     );
     const noFrameOptions = {};
     switch (process.platform) {
@@ -333,35 +218,6 @@ function* forkInitialRendererProcess() {
         },
       },
     );
-    const teamStatusListWindow = yield call(
-      windowsManagerSagas.forkNewWindow,
-      {
-        url: teamStatusListUrl,
-        scopes: ['mainRenderer'],
-        BrowserWindow,
-        options: {
-          show: false,
-          width: 320,
-          height: 350,
-          minWidth: 320,
-          minHeight: 350,
-          fullscreenable: false,
-          resizable: false,
-          transparent: true,
-          skipTaskbar: true,
-          alwaysOnTop: true,
-          frame: false,
-          webPreferences: {
-            devTools: (
-              process.env.NODE_ENV === 'development'
-              || process.env.DEBUG_PROD === 'true'
-            ),
-            webSecurity: false,
-            backgroundThrottling: false,
-          },
-        },
-      },
-    );
 
     const menuBuilder = new MenuBuilder(win);
     menuBuilder.buildMenu();
@@ -373,55 +229,13 @@ function* forkInitialRendererProcess() {
       ],
     });
 
-    const teamStatusListCloseChannel = windowsManagerSagas.createWindowChannel({
-      win: teamStatusListWindow,
-      events: [
-        'close',
-      ],
-    });
-
-    const teamStatusListToggleChannel = windowsManagerSagas.createWindowChannel({
-      win: global.teamStatusListTray,
-      events: [
-        'click',
-      ],
-    });
-
-    const blurChannel = windowsManagerSagas.createWindowChannel({
-      win: teamStatusListWindow,
-      events: [
-        'blur',
-      ],
-    });
-
-    yield fork(onToggleClick, {
-      win: teamStatusListWindow,
-      channel: teamStatusListToggleChannel,
-    });
-
-    yield fork(onBlur, {
-      win: teamStatusListWindow,
-      channel: blurChannel,
-    });
-
     yield fork(onClose, {
       win,
       channel: closeChannel,
     });
 
-    yield fork(onClose, {
-      win: teamStatusListWindow,
-      channel: teamStatusListCloseChannel,
-    });
-
     const eventsChannel = windowsManagerSagas.createWindowChannel({
       win,
-      events: browserWindowInstanceEvents,
-      webContentsEvents: webContentsInstanceEvents,
-    });
-
-    const teamStatusListEventsChannel = windowsManagerSagas.createWindowChannel({
-      win: teamStatusListWindow,
       events: browserWindowInstanceEvents,
       webContentsEvents: webContentsInstanceEvents,
     });
@@ -432,11 +246,6 @@ function* forkInitialRendererProcess() {
       scope: 'all',
     });
 
-    yield fork(windowsManagerSagas.onEventLogger, {
-      windowId: teamStatusListWindow.id,
-      channel: teamStatusListEventsChannel,
-      scope: 'all',
-    });
   } catch (err) {
     console.log(err);
   }
